# backend/requirement/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: AI Task Dashboard API
  description: >
    Automatically derived CRUD API documentation based on `Data requirements.md`.
    All responses follow the standardized API response envelope.
  version: 1.1.0
servers:
  - url: /api/v1
    description: Main API Server
...
paths:
...
  /challenges:
    get:
      tags: [Challenges]
      summary: Get all challenges
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: A list of challenges including derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          challenges:
                            type: array
                            items:
                              $ref: '#/components/schemas/ChallengeResponse'
    post:
      tags: [Challenges]
      summary: Create a challenge
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeBase'
      responses:
        '201':
          description: Challenge created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          challenge:
                            $ref: '#/components/schemas/ChallengeResponse'
...
  /challenges/{virtualId}:
    get:
      tags: [Challenges]
      summary: Get challenge by Virtual ID (e.g., CH-001)
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the challenge
      responses:
        '200':
          description: Challenge object with derived fields
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          challenge:
                            $ref: '#/components/schemas/ChallengeResponse'
    put:
      tags: [Challenges]
      summary: Update a challenge
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeBase'
      responses:
        '200':
          description: Challenge updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          challenge:
                            $ref: '#/components/schemas/ChallengeResponse'
    patch:
      tags: [Challenges]
      summary: Update challenge status (Swim Lane)
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
      description: >
        Update the swim lane status of a challenge.
        **Constraint:** Once a challenge is in `pilot` status it cannot be moved back to `submitted`.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [submitted, ideation, pilot, completed, archive]
                  description: New swim lane status. Transition from 'pilot' back to 'submitted' is forbidden.
      responses:
        '200':
          description: Challenge status updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          challenge:
                            $ref: '#/components/schemas/ChallengeResponse'
        '400':
          description: Bad Request - Invalid status transition (e.g., pilot to submitted is not allowed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
    delete:
      tags: [Challenges]
      summary: Delete a challenge
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Challenge deleted
  /challenges/{virtualId}/upvote:
    post:
      tags: [Challenges]
      summary: Toggle upvote for a challenge
      operationId: toggleChallengeUpvote
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the challenge
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Upvote toggled successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          challenge:
                            $ref: '#/components/schemas/ChallengeResponse'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /challenges/{virtualId}/subscribe:
    post:
      tags: [Challenges]
      summary: Toggle subscription for a challenge
      operationId: toggleChallengeSubscribe
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the challenge
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Subscription toggled successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          challenge:
                            $ref: '#/components/schemas/ChallengeResponse'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'

  # ==============================
  # IDEAS ENDPOINTS
  # ==============================
...
  /comments/challenge/{virtualId}:
    get:
      tags: [Comments]
      summary: Get comments for a challenge by virtualId
      operationId: getCommentsByChallengeVirtualId
      description: >
        Fetches all comments for a challenge. The virtualId prefix (CH-) maps to
        Comment Type "CH". Internally resolves the challenge by virtualId and
        queries comments where type="CH" and typeId matches.
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the challenge (e.g., CH-001)
      responses:
        '200':
          description: List of comments for the challenge
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comments:
                            type: array
                            items:
                              $ref: '#/components/schemas/CommentResponse'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
  /comments/challenge/{virtualId}/count:
    get:
      tags: [Comments]
      summary: Get comment count for a challenge by virtualId
      operationId: getCommentCountByChallengeVirtualId
      parameters:
        - name: virtualId
          in: path
          required: true
          schema:
            type: string
          description: The virtual ID of the challenge (e.g., CH-001)
      responses:
        '200':
          description: Comment count for the challenge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountApiResponse'
        '404':
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseEnvelope'
...
components:
  parameters:
...
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
      description: Amount of records to return for pagination
    OffsetParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
      description: Amount of records to skip for pagination
  schemas:
...
    ApiResponseEnvelope:
      type: object
      required: [status, message, timestamp]
      properties:
        status:
          type: string
          enum: [success, failed]
          description: Indicates whether the request was successful or failed
        message:
          type: string
          description: Human-readable success or failure message
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 date-time of the response
        data:
          type: object
          nullable: true
          description: Domain-specific payload. May be null on failure.
    CountApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                count:
                  type: integer
                  description: The total count

    # ------------------
    # MINIMAL SCHEMAS (FOR REUSE)
    # ------------------
    UserMinimal:
      type: object
      required: [_id, name]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        name:
          type: string
        email:
          type: string
        companyTechRole:
          type: string
        role:
          type: string
          enum: ['ADMIN', 'MEMBER', 'USER']
...
    IdeaMinimal:
      type: object
      required: [_id, title]
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        ideaId:
          type: string
          description: ID-XXXX format
        title:
          type: string
    CommentMinimal:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB hex ID
        comment:
          type: string
        userId:
          type: string
        createdat:
          type: string
          format: date-time

    # ------------------
    # BASE DB SCHEMAS mapping to input DB fields
    # ------------------
...
    ChallengeBase:
      type: object
      properties:
        title:
          type: string
        opco:
          type: string
        platform:
          type: string
        description:
          type: string
        summary:
          type: string
        outcome:
          type: string
        timeline:
          type: string
        portfolioLane:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
          default: ["ai"]
        constraint:
          type: string
        stakeHolder:
          type: string
        virtualId:
          type: string
        status:
          type: string
        userId:
          type: string
        month:
          type: integer
        year:
          type: integer
        upVotes:
          type: array
          items:
            type: string
        subcriptions:
          type: array
          items:
            type: string
        viewCount:
          type: integer
        timestampOfStatusChangedToPilot:
          type: string
          format: date-time
          nullable: true
        timestampOfCompleted:
          type: string
          format: date-time
          nullable: true
...
    CommentBase:
      type: object
      properties:
        userId:
          type: string
        comment:
          type: string
        type:
          type: string
          enum: ['CH', 'ID']
        typeId:
          type: string
        createdat:
          type: string
          format: date-time
...
    ChallengeResponse:
      allOf:
        - $ref: '#/components/schemas/ChallengeBase'
        - type: object
          properties:
            _id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            countOfIdeas:
              type: integer
              description: "(DERIVED) Total active ideas linked"
            ownerDetails:
              $ref: '#/components/schemas/UserMinimal'
            contributorsDetails:
              type: array
              items:
                $ref: '#/components/schemas/UserMinimal'
              description: "(DERIVED) Populated contributors users details"
            commentCount:
              type: integer
              description: "(DERIVED) Calculated from Comments table"
            ideaList:
              type: array
              items:
                $ref: '#/components/schemas/IdeaMinimal'
              description: "(DERIVED) Full list from Idea collection"
            comments:
              type: array
              items:
                $ref: '#/components/schemas/CommentMinimal'
              description: "(DERIVED) Full list from Comment collection"
            contributors:
              type: array
              items:
                $ref: '#/components/schemas/UserMinimal'
              description: "(DERIVED) Owner details of all ideas under challenge"
...
    CommentResponse:
      allOf:
        - $ref: '#/components/schemas/CommentBase'
        - type: object
          properties:
            _id:
              type: string
            userDetails:
              $ref: '#/components/schemas/UserMinimal'
...